Virtualize machine software interrupt (MSI)
This commit implements proper MSI virtualization. For that purpose it
virtualizes the list of pending MSI in the virtual CLINT. When a vCPU
sends an MSI to another core the virtual CLINT triggers a physical MSI
to signal the core that a virtual interrupt should be injected. If a
vCPU triggers an MSI on the current core then a virtual interrupt is
injected directly without causing a physical interrupt.

This is an important first step toward allowing out-of-band (host) MSI,
which are required by some security policies for cross-caches flushes.