Use distinct payload and firmware target
In the past we merged the firmware and payload target files as they are
now identical. However that caused more re-compilation than necessary
because the firmware and the payload use different flags (in particular
linker flags for the start address).

This commit revert that changes and separate the config files for the
payload and firmware. After this change the test time is reduced from
30s to 10s on my machine.