Make runner test path relative to workspace root
This makes it possible to run ` cargo run -- test` from a subfolder,
rather than only from the root of the workspace.
Caveat: ` cargo run -- test` wont work if the current folder is a crate
other than the runner (for instance in `./src`), as this will invoke run
for the current crate instead of the runner.