Add Spike tests to the CI
This commit installs the Spike emulator in the CI (using our own
pre-built images, as Spike is not available in apt) and runs the Spike
tests.

The commit adds a `--strict` flag which causes the tests to fail is some
of them are skipped. Tests can be skipped if an emulator is missing, so
this will catch errors in the Spike installation rather than accept
broken tests.