Runner: enable passing "--strict" flag in env variables
With this change the `--strict` flag can be enabled by passing the
'MIRALIS_RUNNER_STRICT' environment variable. This way we can set it
using the environment variable in the CI and remove it from the
justfile, which will make the test pass even if Spike is not installed
(with some warnings).