Runner: add support for Kani
This commit adds support for model checking with Kani. This is done from
the runner so that we can simply ignore the model checking if Kani is
not installed and we are not running in strict mode (e.g. in the CI).

The Sail model is **not** included in this commit, so nothing can
actually be checked with this commit.