Add mret model checking
This PR adds an initial model checking of the mret instruction against a
reference extracted from the Sail specification. We initialize the
registers with symbolic values and emulate an mret instruction in both
Miralis and the Sail model. If the two returns a virtual context with
the same symbolic values, then the verification is successful.

It is also possible to run the model checking without Kani as a simple
test, in this case the values are not symbolic, but rather arbitrary
(fixed) values returned by Default::default().


Co-authored-by: François Costa <frankost.costa@gmail.com>