Skip write with CSRRS and CSRRS with x0 as bitmask
This change makes it easier to support pseudo-instructions such as
RDTIME which gets translated into CSRRS with x0 as the mask. Currently
we don't handle writes to `time` for instance, but we previous to this
patch we were seeing writes (of an idential value) because CSRRS
emulation always goes through the write phase.