Bugfix: lost interrupts in WFI and MIE
This commit fixes two related bugs that caused interrupts to be lost in
some circumstances.
First while executing a physical WFI the hardware was not properly
re-configured to enable the desired interrupts, and thus sometimes did
not wake-up.
Secondly The MIE CSR was not always properly configured and sometimes
disabled some interrupts who should not be (because they are now used by
Miralis for things such as sending a virtual MSI across cores).

With those two fixes in multi-core Linux works again as it should,
waking the two harts appropriately.