Runner: add option for expected output in tests
This commit adds a new test parameter, 'expect', which takes a substring
that should be present in the output of the test. This is useful to
ensure certain things happen. For instance Linux can successfully boot
on a multi-core machine even if a hart is not responding, which this
patch we can check that all harts booted properly.