Platform: make virtual device array variable size
Prior to this commit the size of the array of virtual devices was fixed
to 2, which in turn implies that at least 2 PMP slots are consumed by
the devices
In preparation for adding a PLIC virtual device, this commit makes the
size of the array variable. That way we can keep the test device on
PMP-rich platforms such as QEMU while keeping the bare minimum (PLIC and
CLINT) on the VisionFive board.