Add virtual PLIC device
This commit adds a simple PLIC virtual device, which performs some
checks and logging before transparently writing to the physical PLIC.
To support more virtual devices and drivers, this commit refactors
existing code, enabling support for a variable number of devices on each
platform, and a cleaner code organisation into modules.

It turns out the PLIC can not be efficiently virtualized without
consuming too much PMP entries (i.e. a number that scales with the number
of harts), as the MMIO registers for the supervisor and machine mode are
alternating: we can't simply trap on the machine-mode registers without
also trapping all the supervisor mode access.
For this reason the virtual PLIC is not yet used by Miralis by default,
but can still be useful for debugging purposes and tracing firmware
accesses.