Bump ubuntu-latest to 24.04
This commit bumps the ubuntu version used in the CI. Currently
`ubuntu-latest` points to ubuntu 22.04, but there is an ongoing roll-out
of version 24.04 until Jan 17. See:

https://github.com/actions/runner-images/issues/10636

The new version should have an updated QEMU version which fixes bugs
around PMPs. Rather than waiting month more we simply bump the
ubuntu version manually, but we can revert the change back to
`ubuntu-latest` after Jan 17.

In addition mark the workflow with `workflow_dispatch` to allow manually
triggering the tests, which is helpful for debugging the CI.