Disable support for machine external interrupts (MEI)
This commit disables support for machine-mode external interrupt. As we
currently don't emulate the PLIC (which would require trapping on M-mode
accesses too) we don't have a proper solution to handle MEI (which would
require claiming the interrupt, hence interacting with the physical
PLIC an potentially racing with the OS/firmware).
So instead we simply decide to disable MEI as OpenSBI doesn't use them.
In the future we could re-enable them selectively if the PLIC is
virtualized, to support firmwares such as Zephyr with MEI enabled.