Push world_switch cost for the statistics of the website
This commit allows to track the latency cost of a firmware trap and
world switch. The statistics are then pushed to a separate repo for
display on the Miralis website's dashboard.
This commit also adds a new test to ensure the configuration we use to
benchmark Miralis is always working.


Co-authored-by: Charly Castes <charly.castes@epfl.ch>