Enable supported MISA extension
We initially disabled most of the misa extensions when we were still
working on the basics of Miralis. Now that Miralis is maturing we do
support most of the extensions (with the special case of user-level
interrupts, the 'N' extension and floating points until we handle
mstatus.FS). This commit marks most extensions as enabled.