Model checking: set nb_pmp to 64 by default
We used to have the number of PMP fixed to 16, but this makes the PMP
write check fail because in the Sail model the number of PMPs are
checked at the decoding step, not when accessed.

Increasing the number of PMPs to 64 makes the verification of the PMPs
much longer (1600s on my machine). So we might want to find a better
solution in the longer term.