PerformanceFix: Split the decoder into two components.
Currently, we have a single decoder. In reality, loads, stores, and
illegal instruction traps handle a disjoint set of instructions. By
splitting the decoder, we can save a few cycles.

Note: illegal instructions can also be caused by floating points or
vector extensions while disabled. We don't check for this currently, but
we should in the future.


Co-authored-by: Charly Castes <charly.castes@epfl.ch>