Add Support for the Premier P550 Board
This commit introduces support for the Premier P550 board, a
high-performance RISC-V platform from SiFive. Compared to the
VisionFive2, the P550 offers significantly more computational power and
enhanced feature support. Currently, booting Miralis on this board
results in a kernel panic with the message: "not yet implemented:
Virtualize machine external interrupt." Further work is required to
address this issue.

To boot OpenSBI/U-Boot on the board we load Miralis in between Open-SBI
and U-Boot, at 0x80080000. With this configuration we can run unmodified
OpenSBI and U-Boot as all software keep their default location.

Co-authored-by: Charly Castes <charly.castes@epfl.ch>