Verification: check exception virtualizaiton
Currently, we only verify formally the correct virtualization of
interrupts, this commit adds the verification of exception
virtualization.

In Miralis the virtualization of exception relies on values in
`ctx.trap_info`, which is filled in the Miralis trap handler. In order
to supply the trap information we emulate an exception in Sail and
initialize the `ctx.trap_info` with the resulting values. We then
compare the emulation of the exception in Miralis with another Sail
emulation.


Co-authored-by: Charly Castes <charly.castes@epfl.ch>