Counter benchmark per Mcause exception
This commit introduces a new benchmark which tracks the number of interrupts for each Mcause exceptions. This benchmark is used only for debugging and doesn't return the values to the kernel when there is an ecall request for measurements. Instead, it simply prints then to UART such that we can analyse the output and have a clearer understanding of the behavior in the system.