Introduction of a precondition verifier in illegal instruction emulation.
Currently, we run the formal verification with a set of precondition on it, and we assume that those conditions holds in Miralis. For example, we say that the register 0 is always zero in the virtual context or that we have support for Userspace and Supervisor mode in the misa register. This commit add the equivalent assertions and introduces them in Miralis to detect any violation of the preconditions. This commit is a first step in the direction.