Refactor Default Policy
Currently, some functions implement a default behavior in the Policy trait but are not specified in the default policy. This commit adds a default behavior for all functions in the Policy trait and provides a minimal implementation of the default policy with a constructor and a name. This refactoring makes the code cleaner, as other policies can benefit from default behaviors in the future.