Introducing test (eid, fid) pair in check_nb_registers_to_forward_per_eid_fid
Currently, there is no SBI rule requiring 6 registers as input, excluding eid and fid. To facilitate testing, a new filtering rule has been added, allowing the 6-register flow to pass through the protect payload policy. It’s important to note that this change is not part of the standard and is intended solely for testing purposes.