Add support for custom CSRs
The P550 implements a few custom registers to configure various
properties of the micro-architecture. As custom registers varies across
implementation we need a mechanism to deal with them. This commit adds
support for for managing custom registers on a per-platform basis.

One of the main annoyance is that the CSR ID is part of the assembly
instruction, rather than being passed as an operand. This means that
each custom CSR needs its own instruction. To avoid bloating Miralis
with unneeded instructions we provide a new macro to instantiate `csrr`
and `csrw` within the platform code.

On the P550 we simply forward the read/writes of custom CSRs. Those CSRs
can affect the system security, especially vulnerability to transient
execution attacks as they can enable micro-architectural optimisation
that lead to more speculative execution. Note that a custom policy could
update those CSRs when switching to a sensitive security context.