Bugfix: compressed loads and stores have the wrong offset
This commit fixes a bug in the decoding of compressed load and stores.
In short the encoded immediate has bits [6:2] for words and [7:3] for
double words, which means that the shift by word/double word size is
already encoded. However the code added a second manual shift, hence
producing erroneous values.


Co-authored-by: Charly Castes <charly.castes@epfl.ch>