Add ```todo!``` statement for decoding stack-based loads and stores
Currently, if we face a stack-based load or store we will silently pass it as a unknown instruction. In the case of the protect payload policy, this will simply lead to a silent crash. This commit adds the todo statement such that if this instruction appears, we get notified we must implement it and the program halts execution