Introduce offload policy and multiplex timer interrupts in Miralis
This commit introduces a new policy. This policy is a bit different from the previous one in the sense that the main goal is the performance here. The previous measurements of Miralis showed that a huge amount of traps came from three sources. Reading the time register, setting the timers for the payload and emulating misaligned load and stores. This policy offloads the three points and handles the logic in Miralis.