Add support for supervisor timer offload
This commit adds support for the supervisor timer SBI into the offload
policy. It enables emulation of the supervisor timer (i.e. Sstc)
functionalities in Miralis rather than by relying on the firmware.
This significantly increases performance on boards without the Sstc
extension.

The implementation relies on a virtualizaiton of the physical timer
deadline, where two possible deadlines (the one from the firmware, and
the one for the payload) are multiplexed into the physical `mtimecmp`
register.


Co-authored-by: Charly Castes <charly.castes@epfl.ch>