Introduce offload policy
This commit introduces a new policy. This policy is a bit different from
the previous one in the sense that the main goal is the performance
here. The previous measurements of Miralis showed that a huge amount of
traps came from three sources. Reading the time register, setting the
timers for the payload and emulating misaligned load and stores.

This commit adds direct support for misaligned load/store and time
register read emulation in the offload policy, significantly improving
performance on boards without hardware support for those features.
The timer offload will come in a later commit.


Co-authored-by: Charly Castes <charly.castes@epfl.ch>