Bugfix: sbi filtering rules doesn't take into account the legacy extension
The script we have to filter out the registers used by the SBI didn't parse the legacy extension correctly. This update fixes the issues. One may notice that some filtering rules changes in the middle. This is simply due to the fact that we used a newer version of the SBI and is not a bug.