Split Instr in Load, Store, and Illegal instr
This commit splits the generic Instr enum in three components:
IllegalInstr, LoadInstr, and StoreInstr. The instruction decoder is also
split accordingly.

Such a decomposition makes sense as the `mcause` CSR indicates the kind
of instruction that caused a trap, therefore the right decoded can be
called based on the error code. This saves works in the decoder itself,
and makes the overall trap handling faster.

As a downside this Miralis no longer has a generic instruction decoder,
and must therefore always rely on the `mcause` code to determine which
kind of instruction caused a fault.


Co-authored-by: Charly Castes <charly.castes@epfl.ch>