Bugfix: Fallback to load instructions with MPRV
Commit ee6279391358bccf39999600e09694498db067fb introduces a new bug.
When we emulate the misaligned load and stores, we don't copy anymore
from the previous mode. If the given address is virtual, we loose the
mapping and therefore the code will be wrong. This commit patches the
bug.

This commit fixes changes the fallback case of `get_raw_faulting_instr`
to use MPRV to load the instruction. In other words, it will copy the
instruction using the previous mode privileges.

Co-authored-by: François Costa <frankost.costa@gmail.com>