Introduce module combinator
This commit introduces a way to combine multiple policy modules into a
single new module. We rely on a proc-macro to generate only required
code at compile time, which is needed to ensure proper performance.
We centralise the use of the macro into a single file, and made it so
that adding a new module does not require touching any macro code.

This commit changes the way modules are added to a configuration, which
now uses the a `modules` table. In the future, we will unify other
functionalities, such as benchmarks, under the module field.

```toml
[modules]
modules = ["offload", "protect_payload"]
```

IMPORTANT NOTE: the ordering of the modules matter! They will be called
in the order defined in the config file.