Add support for measuring exceptions by categories in  `counter benchmark`
Before this commit, the `counter benchmark` measured the number of world
switches and firmware trap. This commit refines a bit the granularity by
adding a few other categories such as request for time, set time, IPI
broadcast, remote fence and misaligned instruction.

NOTE: because of the slight refactor of the benchmark interface this
commit deprecates the old `default` benchmark. As the benchmark is no
longer accurate we remove the corresponding tests.