Introduce `boot` benchmark
This benchmark is special and is used to measure the behavior of the
linux kernel while it is booting. It will count the number of exception
per category (non-offloadable, misaligned, read time, set timer,
broadcast ipi and remote fences). After some point the benchmark simply
halts the execution of Miralis.