Add building miralis for the `visionfive2` and `premierp550` boards in the test pipeline
We won't detect all build errors in the `visionfive2` and `premierp550`
in the platform at the moment. This commit adds them in the pipeline to
be sure they are buildable after every commit.

Note: this change significantly increases the test time. This is because
the builds for the board trigger a re-compilation of all dependencies.
It is not clear if such a re-compilation is needed, maybe it can be
optimized away in the future.


Co-authored-by: Charly Castes <charly.castes@epfl.ch>