PerformanceFix: Make the physical clint driver stateless
The physical clint driver should be a simple writer and there is no reason to protect it with a lock. (OpenSBI doesn't protect it with a lock as well.) This commit removes the overhead and refactors at the same time the code around the clint driver to make it easier to understand.