Remove lock around ClintDriver
We were measuring a non-negligible performance impact of the lock around
the ClintDriver, the offload policy rely on the driver to emulate reads
to mtime.
This commit removes the lock around the ClintDriver to lower contention,
and we do measure a significant improvement (Miralis with offload policy
is now as fast as native for most workloads).

Of course the lock was there for a reason, we basically deferred the
reflection of how to access concurrent reads/write on the CLINT. It
turns out Rust does not have clear guidelines regarding MMIO
read/writes, so we aligned our implementation on the OpenSBI C version
and use racy accesses. Note that there is no cache effects for MMIO,
that we use volatile reads and do never materialize a `&` or `&mut`
reference to the CLINT, only raw pointers. This should keep us on the
safe side, although we should keep-up with the Rust memory model for
MMIO as it evolves.