Introduce function to redirect traps to the S-mode payload
This commit adds an OpenSBI function to redirect traps in Miralis directly. We want to use it to forward store, load, and instruction page faults directly to the S-mode payload without going through OpenSBI. Additionally, this function is needed to forward failures from misaligned loads and stores to the payload as well.