Bugfix: correct NB_VIRT_DEVICES on VisionFive 2
We removed the test device for the VisionFive 2 platform, but did not
update properly the NB_VIRT_DEVICES constant on the Platform trait.
Furtunately the issue was caught by an assertion.
This commit fixes the issue by setting NB_VIRT_DEVICES to be the len of
the VIRT_DEVICES constant, which is possible since slice::len is a const
fn.