Mirgrate benchmarks to the new module system
We used to have a concept of "benchmark modules", which was useful to
have both a policy and a benchmark installed. However, with the new
module system and the ability to combine multiple modules there is no
such need for a distinction anymore. This commit migrates the benchmarks
to the new module system, simplifying the code base and reducing the
number of concepts.