Build Miralis in release mode on the Premier P550
At the difference of the VisionFive 2, we only have one config file for
the Premier P550, so it makes sense to change it for building a release
image.

Another option would be to create a separate release configuration, but
I would like to avoid config file proliferation. We already have a bit
too many configuration files for my taste, so let's not create more for
now and let's see if we can clean that up in the future.