Only display boot info on the boot hart
Previously we were displaying the boot info on all harts, which was just
generating lots of redundant logs on our hardware platforms (with 4 and
5 harts). This commit reduces the clutter by not logging redundant
information.

We don't synchronize harts yet (except for .bss initialization). We
should probably add a barrier to let the boot hart perform potential
one-time initialization first. That would require testing on the
VisionFive 2, however, as we had some interesting bugs with hart ID
numbering there.