Remove duplicate configuration env vars
The configuration environment variables were duplicated in two places:
in the runner which sets those variables, and in Miralis which reads
them. This commit centralizes the definitions of the environment
variables to minimize the risk of error when adding or updating
configuration variables. We also had duplicate environment variables in
a few other places, which are removed too.