Upgrade toolchain
With the recent migration toward softcore-rs and the removal of the old
Sail mode, we can finally upgrade the toolchain. This was held back by
the dependency of the old model on the now broken const generic expr
feature.

This will open the door to further upgrade, such as migrating to Rust
edition 2024.