Update `mstatus` handling to match the Sail spec
This commit simplifies the `mstatus` handling and update the supported
fields to match the newer version of the Sail specification. This should
not impact the behavior of Miralis on supported platforms.

NOTE: it seems we do not write the PMP to the hardware when MPRV
changes. That looks like a bug that we need to investigate in the
future.