Fix PMP virtualization corner case on partial matching
It turns out that partially matching a PMP entry always result in a trap
irrespective of the LRWX bits and privilege mode. This means that an
M-mode software can trap when a memory access partially matches a PMP
entry even if the lock bit is not set.
However, if the PMP is OFF there is no concept of partial match, be
cause the PMP entry matches no addresses. Because Miralis was
virtualizing PMP entries by turning them OFF if the lock bit is not set,
this resulted in a divergence of behavior on partial PMP match.

Therefore, to properly virtualize PMP this commit changes Miralis to
instead keep the virtual PMP addresses and matching mode, but forces the
access rights to be RWX when the lock bit is not set. This ensures that
the virtualized firmware keep accesses to all memory, but makes it trap
on partial matches, ensuring the same behavior as native M-mode code.

This commit also update the PMP verification to ensure the virtualized
firmware executes as if it was running on a reference physical machine.