Add support for testing with random values in model checking
When running standard Rust tests we use default values instead of
symbolic values in the model checking crate. In practice most of these
values are 0.

This commit adds a new "rand" feature to the `model_checking` crate
which instead uses random value. This is useful finding failure cases.
In the future we might want to support proper fuzzing, but this will
require a major refactor of the model checking code. This change, in
contrast, is pretty self contains and already find errors.

How to use:

```
cargo test -p model_checking --features rand <test_fn_name>
```