Bugfix: Test device covers memory used by payload
In 550ace64 we increased the size of the test device to account for the
PMP granularity, from 0x8 to 0x10000. However with this change the test
device now covers memory that happens to be used by some payload,
notably U-Boot.

This commit re-locates the test device next to the CLINT, where it
should not be accessed by firmware or payload.