Bugfix: tracing trap handler loses 1 bit in register
The `tracing_trap_handler` used to temporary store a register (x5) into
mepc, while incrementing mepc to skip the trapping instruction. It turns
out mepc has its bit 0 (and bit 1 if the C extension is not supported)
hardwired to 0, so we were losing bit 0 of x5.
The new implementation stores x5 directly in mscratch to prevent losing
any bits.

This bug was found using a new tool to convert assembly to Rust code,
building on softcore-rs, so I think this is a good indication that this
new tool is useful, and I hope to integrate it more deeply into Miralis
in the future :)