Switch to softcore-asm to support user-space tests
Miralis previously had a mock implementation of hardware manipulation
functions, which reflected the changes into an emulated core provided by
softcore-rs. This was a great first step for verification, but we
couldn't extend verification to the actual assembly snippets, which are
one of the most tricky part to get right.
With this change we remove the mock hardware implementation, and instead
use the new softcore-asm crate which provides a macro that emulates
`asm`. We now replace all instances of `core::arch::asm` with a custom
`soft_asm` macro that either invokes `core::arch::asm` or
`softcore_asm_rv64::asm` when compiling to user-space. That way, the
same code is tested when running on hardware or during unit tests and
verification.