Bugfix: protect Keystone enclave when switching to firmware
When the enclave traps to firmware the Keystone policy must lock the
enclave before transferring control to the firmware. This commit
interposes on the `switch_from_payload_to_firmware` to check if the
payload is an enclave, and if it is it locks the enclave before
returning to firmware. On the way back, the policy also check if an
enclave is running to unlock it again when necessary.

Fixes #477